fs: Change max inode quantity in dfs to 255 (uint8_t)y

+++ src/fs/driver/dfs/dfs.h
@@ -26,8 +26,8 @@
 /* Non-VFS declarations */
 struct dfs_sb_info {
 	uint8_t magic[2];
-	uint32_t inode_count;
-	uint32_t max_inode_count;
+	uint8_t inode_count;
+	uint8_t max_inode_count;
 	uint32_t buff_bk;         /* For buffer-based writing */
 	uint32_t free_space;
 };
